requirements.txt
git asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-embed-video==1.4.10
idna==3.10
mysqlclient==2.2.4
pillow==11.0.0
python-decouple==3.8
razorpay==1.4.2
requests==2.32.3
setuptools==75.3.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
